+++ Assets/ParticleThumbnail&Preview/Editor/AssemblyInfo.cs
@@ -1,0 +1,5 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("ParticleThumbnailAndPreview.Editor.URP")]
+[assembly: InternalsVisibleTo("ParticleThumbnailAndPreview.Editor.HDRP")]
+[assembly: InternalsVisibleTo("ParticleThumbnailAndPreview.Editor.Tests")]